consumer-imagecontext/requirements.txt
flask
pika
transformers
torch
Pillow
numpy
